Fix bitpacking edge case behaviors.
The current bitpacking code does very silly things when dealing with
little endian and signed numbers. This commit aims to fix both and add a
few tests to help ensure ongoing correctness of behaviors.

This change is somewhat breaking, but the old behavior is so broken,
that it is pretty doubtful anyone was actually relying on it.

Unfortunately, this change may also introduce performance regressions,
as it introduces changes in the hotpath of bitpacking decoding and
encoding and integer decoding in general. Although Restruct is not
currently designed for high throughput, this is something that should be
investigated as soon as possible.

Fixes #46.